refactor(repository): refactor property management by consolidating repositories and enhancing CRUD operations
