+++ apps/api/src/services/property-crud.service.ts
@@ -5,15 +5,18 @@ import type {
 import { prisma } from "../configs/prisma.config.js";
 import { nanoid } from "nanoid";
 import slugify from "@sindresorhus/slugify";
-import { PropertyRepository } from "../repositories/property.repository.js";
+import { PropertyCrudRepository } from "../repositories/property-crud.repository.js";
+import { PropertyRelationsRepository } from "../repositories/property-relations.repository.js";
 import { PropertyValidator } from "../utils/property-validator.js";
 import { mapFacilities, mapPictures, mapRooms } from "../utils/mappers.js";
 
 export class PropertyCrudService {
-  private repository: PropertyRepository;
+  private crudRepository: PropertyCrudRepository;
+  private relationsRepository: PropertyRelationsRepository;
 
   constructor() {
-    this.repository = new PropertyRepository();
+    this.crudRepository = new PropertyCrudRepository();
+    this.relationsRepository = new PropertyRelationsRepository();
   }
 
   async createProperty(data: CreatePropertyInput) {
@@ -38,12 +41,17 @@ export class PropertyCrudService {
           address: data.address,
           latitude: data.latitude ?? null,
           longitude: data.longitude ?? null,
+          maxGuests:
+            (data.rooms || []).reduce(
+              (sum: number, r: any) => sum + (r.capacity ?? 1),
+              0
+            ) || 1,
           Facilities: mapFacilities(data.facilities),
           Pictures: mapPictures(data.pictures),
           Rooms: mapRooms(data.rooms),
         };
 
-        return this.repository.create(propertyData);
+        return this.crudRepository.create(propertyData);
       },
       {
         maxWait: 15000,
@@ -57,10 +65,12 @@ export class PropertyCrudService {
   async getPropertyBySlug(slug: string) {
     PropertyValidator.validateSlug(slug);
 
-    const property = await this.repository.findUniqueBySlug(slug);
+    const property = await this.crudRepository.findUniqueBySlug(slug);
     PropertyValidator.validatePropertyExists(property);
 
-    const reviewStats = await this.repository.getReviewStats(property.id);
+    const reviewStats = await this.relationsRepository.getReviewStats(
+      property.id
+    );
 
     return {
       ...property,
@@ -72,7 +82,10 @@ export class PropertyCrudService {
   async getPropertyById(propertyId: string, tenantId?: string) {
     PropertyValidator.validatePropertyId(propertyId);
 
-    const property = await this.repository.findUniqueById(propertyId, tenantId);
+    const property = await this.crudRepository.findUniqueById(
+      propertyId,
+      tenantId
+    );
     PropertyValidator.validatePropertyExists(property, tenantId);
 
     return property;
@@ -81,25 +94,25 @@ export class PropertyCrudService {
   async getPropertiesByTenant(tenantId: string) {
     PropertyValidator.validateTenantAccess(tenantId);
 
-    return this.repository.findManyByTenant(tenantId);
+    return this.crudRepository.findManyByTenant(tenantId);
   }
 
   async deleteProperty(propertyId: string, tenantId: string) {
     PropertyValidator.validatePropertyId(propertyId);
     PropertyValidator.validateTenantAccess(tenantId);
 
-    const property = await this.repository.findFirstByIdAndTenant(
+    const property = await this.crudRepository.findFirstByIdAndTenant(
       propertyId,
       tenantId
     );
     PropertyValidator.validateDeletePermission(property);
 
-    const activeBookingCount = await this.repository.getActiveBookingCount(
+    const activeBookingCount = await this.crudRepository.getActiveBookingCount(
       propertyId
     );
     PropertyValidator.validateActiveBookings(activeBookingCount);
 
-    await this.repository.delete(propertyId);
+    await this.crudRepository.delete(propertyId);
 
     return { message: "Property deleted successfully" };
   }
@@ -112,7 +125,7 @@ export class PropertyCrudService {
     PropertyValidator.validatePropertyId(propertyId);
     PropertyValidator.validateTenantAccess(tenantId);
 
-    const existingProperty = await this.repository.findFirstByIdAndTenant(
+    const existingProperty = await this.crudRepository.findFirstByIdAndTenant(
       propertyId,
       tenantId
     );
@@ -133,7 +146,10 @@ export class PropertyCrudService {
           updateData.customCategoryId = data.customCategoryId || null;
         }
 
-        const updated = await this.repository.update(propertyId, updateData);
+        const updated = await this.crudRepository.update(
+          propertyId,
+          updateData
+        );
 
         await this.updateFacilities(propertyId, data);
         await this.updatePictures(propertyId, data);
@@ -169,18 +185,24 @@ export class PropertyCrudService {
     data: UpdatePropertyInput
   ) {
     if ("facilities" in data && data.facilities !== undefined) {
-      await this.repository.deleteFacilities(propertyId);
+      await this.relationsRepository.deleteFacilities(propertyId);
       if (data.facilities.length > 0) {
-        await this.repository.createFacilities(propertyId, data.facilities);
+        await this.relationsRepository.createFacilities(
+          propertyId,
+          data.facilities
+        );
       }
     }
   }
 
   private async updatePictures(propertyId: string, data: UpdatePropertyInput) {
     if ("pictures" in data && data.pictures !== undefined) {
-      await this.repository.deletePictures(propertyId);
+      await this.relationsRepository.deletePictures(propertyId);
       if (data.pictures.length > 0) {
-        await this.repository.createPictures(propertyId, data.pictures);
+        await this.relationsRepository.createPictures(
+          propertyId,
+          data.pictures
+        );
       }
     }
   }
